[fix]: the DrawFrame class is divided into the Image and Frame classes
