+++ src/classes/Frame.py
@@ -1,11 +1,9 @@
-
 from dataclasses import dataclass
 
-
 @dataclass
-class DrawFrame:
+class Frame:
     """
-    Description: Draw frame class for ODT document.
+    Description: Frame class for ODT document.
 
     Attributes:
         _frame_style_name - attribute specifies the style of current frame,
@@ -16,13 +14,7 @@ class DrawFrame:
         _frame_width - attribute specifies the width of the frame,
         _frame_height - attribute specifies the height of the frame,
         _frame_rel_width - attribute specifies height of a drawing object as a relative value within a frame,
-        _frame_rel_height - attribute specifies the width of a drawing object as a relative value within a frame,
-        _image_href - attribute specifies the location of an embedded object,
-        _image_type - attribute always has the value simple in OpenDocument document instances,
-        _image_show - attribute is used to communicate the desired presentation of the ending resource on traversal
-            from the starting resource,
-        _image_actuate - attribute is used to communicate the desired timing of traversal from the starting resource
-            to the ending resource.
+        _frame_rel_height - attribute specifies the width of a drawing object as a relative value within a frame.
     """
 
     _frame_style_name: str
@@ -34,10 +26,6 @@ class DrawFrame:
     _frame_height: float
     _frame_rel_width: str
     _frame_rel_height: str
-    _image_href: str
-    _image_type: str
-    _image_show: str
-    _image_actuate: str
 
     @property
     def frame_style_name(self):
@@ -110,35 +98,3 @@ class DrawFrame:
     @frame_rel_height.setter
     def frame_rel_height(self, value):
         self._frame_rel_height = value
-
-    @property
-    def image_href(self):
-        return self._image_href
-
-    @image_href.setter
-    def image_href(self, value):
-        self._image_href = value
-
-    @property
-    def image_type(self):
-        return self._image_type
-
-    @image_type.setter
-    def image_type(self, value):
-        self._image_type = value
-
-    @property
-    def image_show(self):
-        return self._image_show
-
-    @image_show.setter
-    def image_show(self, value):
-        self._image_show = value
-
-    @property
-    def image_actuate(self):
-        return self._image_actuate
-
-    @image_actuate.setter
-    def image_actuate(self, value):
-        self._image_actuate = value
