+++ src/classes/Paragraph.py
@@ -1,8 +1,6 @@
 import re
 from dataclasses import dataclass, field
-
 from src.helpers.enums import AlignmentEnum
-
 
 @dataclass
 class Paragraph:
