+++ src/classes/TableColumn.py
@@ -1,6 +1,4 @@
-
 from dataclasses import dataclass
-
 
 @dataclass
 class TableColumn:
